Move SavedData and ValidatedData interfaces from internal types
