--- src/types-internal/block-data.d.ts
@@ -1,20 +1,0 @@
-import {BlockToolData} from '../../types/tools';
-
-/**
- * Tool's saved data
- */
-export interface SavedData {
-    tool: string;
-    data: BlockToolData;
-    time: number;
-}
-
-/**
- * Tool's data after validation
- */
-export interface ValidatedData {
-    tool?: string;
-    data?: BlockToolData;
-    time?: number;
-    isValid: boolean;
-}
